Added signal analyzer settings generalization
Updated with definition of SiganSetings object and available classes. USRP and Keysight classes may be incomplete. Updated example with added fields.

This is a breaking change to the spec, since it removes the SensorAnnotation keys `gain_setting_sigan` and `attenutaion_setting_sigan`. These values are now recorded under the SiganSettings object.